+++ src/components/CohortBuilder/ParticipantsTableView/index.js
@@ -1,4 +1,4 @@
-import React, { useEffect } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
 import { compose, withState } from 'recompose';
 
@@ -10,9 +10,6 @@ import ParticipantsTable from './ParticipantsTable';
 import TableErrorView from './TableErrorView';
 
 import './index.css';
-import { connect } from 'react-redux';
-import { getUserSets } from 'store/actionCreators/saveSets';
-import { selectUserSets, selectError, selectIsLoading } from 'store/selectors/saveSetsSelectors';
 
 const ParticipantsTableView = ({
   sqon,
@@ -28,116 +25,84 @@ const ParticipantsTableView = ({
   sort,
   setSort,
   loggedInUser,
-  userSaveSets,
-  userSets,
-  egoGroups,
-}) => {
-  useEffect(() => {
-    if (loggedInUser) {
-      userSaveSets(loggedInUser.egoId);
-    }
-  }, [userSaveSets, loggedInUser]);
-
-  return (
-    <QueriesResolver
-      name="GQL_PARTICIPANTS_TABLE"
-      api={api}
-      queries={[participantsQuery(sqon, sort, pageSize, pageIndex)]}
-    >
-      {({ isLoading, data, error }) => {
-        if (error) {
-          return (
-            <div className="ptv-error-message">
-              <TableErrorView error={error} />
-            </div>
-          );
-        }
-        const isRowSelected = (node) =>
-          allRowsSelected || selectedRows.some((row) => row === node.participantId);
-
-        const dataWithRowSelection = data[0]
-          ? data[0].nodes.map((node) => ({ ...node, selected: isRowSelected(node) }))
-          : [];
-
-        const selectionSQON = selectedRows.length
-          ? {
-              op: 'and',
-              content: [{ op: 'in', content: { field: 'kf_id', value: selectedRows } }],
-            }
-          : sqon;
-
+}) => (
+  <QueriesResolver
+    name="GQL_PARTICIPANTS_TABLE"
+    api={api}
+    queries={[participantsQuery(sqon, sort, pageSize, pageIndex)]}
+  >
+    {({ isLoading, data, error }) => {
+      if (error) {
         return (
-          <ParticipantsTable
-            sqon={selectionSQON}
-            loading={isLoading}
-            data={dataWithRowSelection}
-            api={api}
-            sort={sort}
-            dataTotalCount={data[0] ? data[0].total : 0}
-            downloadName={'participant-table'}
-            onFetchData={({ page, pageSize, sorted }) => {
-              const sorting = sorted
-                .filter((s) => SORTABLE_FIELDS_MAPPING.has(s.id))
-                .map((s) => ({
-                  field: SORTABLE_FIELDS_MAPPING.get(s.id),
-                  order: s.desc ? 'desc' : 'asc',
-                }));
-              setPageIndex(page);
-              setPageSize(pageSize);
-              setSort(sorting);
-            }}
-            onRowSelected={(checked, row) => {
-              const rowId = row.participantId;
-              if (checked) {
-                setSelectedRows((s) => s.concat(rowId));
-                return;
-              }
-              setSelectedRows((s) => s.filter((row) => row !== rowId));
-            }}
-            onAllRowsSelected={(checked) => {
-              // don't keep individual rows selected when "select all" is checked
-              //  to avoid having them selected after "unselect all"
-              setAllRowsSelected(() => checked);
-              setSelectedRows(() => []);
-            }}
-            onClearSelected={() => {
-              setAllRowsSelected(() => false);
-              setSelectedRows(() => []);
-            }}
-            selectedRows={selectedRows}
-            allRowsSelected={allRowsSelected}
-            loggedInUser={loggedInUser}
-            saveSets={userSets.sets}
-            egoGroups={egoGroups}
-          />
+          <div className="ptv-error-message">
+            <TableErrorView error={error} />
+          </div>
         );
-      }}
-    </QueriesResolver>
-  );
-};
+      }
+      const isRowSelected = (node) =>
+        allRowsSelected || selectedRows.some((row) => row === node.participantId);
+
+      const dataWithRowSelection = data[0]
+        ? data[0].nodes.map((node) => ({ ...node, selected: isRowSelected(node) }))
+        : [];
+
+      const selectionSQON = selectedRows.length
+        ? {
+            op: 'and',
+            content: [{ op: 'in', content: { field: 'kf_id', value: selectedRows } }],
+          }
+        : sqon;
+      return (
+        <ParticipantsTable
+          sqon={selectionSQON}
+          loading={isLoading}
+          data={dataWithRowSelection}
+          api={api}
+          sort={sort}
+          dataTotalCount={data[0] ? data[0].total : 0}
+          downloadName={'participant-table'}
+          onFetchData={({ page, pageSize, sorted }) => {
+            const sorting = sorted
+              .filter((s) => SORTABLE_FIELDS_MAPPING.has(s.id))
+              .map((s) => ({
+                field: SORTABLE_FIELDS_MAPPING.get(s.id),
+                order: s.desc ? 'desc' : 'asc',
+              }));
+            setPageIndex(page);
+            setPageSize(pageSize);
+            setSort(sorting);
+          }}
+          onRowSelected={(checked, row) => {
+            const rowId = row.participantId;
+            if (checked) {
+              setSelectedRows((s) => s.concat(rowId));
+              return;
+            }
+            setSelectedRows((s) => s.filter((row) => row !== rowId));
+          }}
+          onAllRowsSelected={(checked) => {
+            // don't keep individual rows selected when "select all" is checked
+            //  to avoid having them selected after "unselect all"
+            setAllRowsSelected(() => checked);
+            setSelectedRows(() => []);
+          }}
+          onClearSelected={() => {
+            setAllRowsSelected(() => false);
+            setSelectedRows(() => []);
+          }}
+          selectedRows={selectedRows}
+          allRowsSelected={allRowsSelected}
+          loggedInUser={loggedInUser}
+        />
+      );
+    }}
+  </QueriesResolver>
+);
 
 ParticipantsTableView.propTypes = {
   sqon: PropTypes.object.isRequired,
   loggedInUser: PropTypes.object,
 };
-
-const mapStateToProps = (state) => ({
-  loggedInUser: state.user.loggedInUser,
-  create: {
-    isLoading: selectIsLoading(state),
-    error: selectError(state),
-  },
-  userSets: {
-    sets: selectUserSets(state),
-    isLoading: false,
-    error: null,
-    isDeleting: false,
-  },
-});
-
-const mapDispatchToProps = (dispatch) => ({
-  userSaveSets: (userId) => dispatch(getUserSets(userId)),
-});
 
 export default compose(
   withApi,
@@ -146,5 +111,4 @@ export default compose(
   withState('selectedRows', 'setSelectedRows', []),
   withState('allRowsSelected', 'setAllRowsSelected', false),
   withState('sort', 'setSort', []),
-  connect(mapStateToProps, mapDispatchToProps),
 )(ParticipantsTableView);