+++ src/components/CohortBuilder/Results.js
@@ -4,111 +4,29 @@ import gql from 'graphql-tag';
 import { compose } from 'recompose';
 import isEmpty from 'lodash/isEmpty';
 import get from 'lodash/get';
-import capitalize from 'lodash/capitalize';
 
 import { injectState } from 'freactal';
-import saveSet from '@kfarranger/components/dist/utils/saveSet';
-
 import Row from 'uikit/Row';
-import { H2 } from 'uikit/Headings';
 import { AppstoreFilled, TableOutlined } from '@ant-design/icons';
-
-import DemographicIcon from 'icons/DemographicIcon';
-import FilesIcon from 'icons/FilesIcon';
-import familyMembers from 'assets/icon-families-grey.svg';
 
 import { Tabs } from 'antd';
 
 import { withApi } from 'services/api';
-import graphql from 'services/arranger';
 
 import TableErrorView from './ParticipantsTableView/TableErrorView';
 import ParticipantsTableView from './ParticipantsTableView';
 import QueriesResolver from './QueriesResolver';
 import EmptyCohortOverlay from './EmptyCohortOverlay';
-import { createFileRepoLink } from './util';
 import Summary from './Summary';
 
-import { Spin, notification } from 'antd';
-import ButtonWithRouter from '../../ui/ButtonWithRouter';
-import { CARDINALITY_PRECISION_THRESHOLD } from '../../common/constants';
-import { roundIntToChosenPowerOfTen } from '../../utils';
+import { CARDINALITY_PRECISION_THRESHOLD } from 'common/constants';
+import Toolbar from './Results/Toolbar/Toolbar';
 
 import './Results.css';
-import { ArrowRightOutlined } from '@ant-design/icons';
 
 const { TabPane } = Tabs;
 const SUMMARY = 'summary';
 const TABLE = 'table';
-const LABELS = {
-  participant: {
-    singular: 'participant',
-    plural: 'participants',
-  },
-  family: {
-    singular: 'family',
-    plural: 'families',
-  },
-  file: {
-    singular: 'file',
-    plural: 'files',
-  },
-};
-
-const formatCountResult = (cardinality, labelKey) => {
-  const isZero = cardinality === 0;
-  if (isZero) {
-    return `No ${capitalize(LABELS[labelKey].singular)}`;
-  }
-
-  const hasMany = cardinality > 1;
-  const label = hasMany
-    ? capitalize(LABELS[labelKey].plural)
-    : capitalize(LABELS[labelKey].singular);
-
-  const isApproximation =
-    cardinality >= CARDINALITY_PRECISION_THRESHOLD && labelKey !== 'participant';
-  if (isApproximation) {
-    const approxSymbol = '\u2248';
-    return `${approxSymbol} ${roundIntToChosenPowerOfTen(cardinality)} ${label}`;
-  }
-  return `${cardinality} ${label}`;
-};
-
-const showErrorNotification = () =>
-  notification.error({
-    message: 'Error',
-    description: 'Unable to create a link to access file repository',
-  });
-
-const generateAllFilesLink = async (user, api, originalSqon) => {
-  try {
-    const fileSet = await saveSet({
-      type: 'participant',
-      sqon: originalSqon || {},
-      userId: user.egoId,
-      path: 'kf_id',
-      api: graphql(api),
-    });
-
-    const setId = get(fileSet, 'data.saveSet.setId');
-
-    return createFileRepoLink({
-      op: 'and',
-      content: [
-        {
-          op: 'in',
-          content: {
-            field: 'participants.kf_id',
-            value: `set_id:${setId}`,
-          },
-        },
-      ],
-    });
-  } catch (e) {
-    showErrorNotification();
-  }
-};
 
 const cohortResultsQuery = (sqon) => ({
   query: gql`
@@ -152,85 +70,30 @@ const Results = ({ activeSqonIndex, sqon = { op: 'and', content: [] }, api, stat
       }
 
       const resultsData = data[0];
+      const isFiltered = !isEmpty(sqon.content);
 
       const participantCount = get(resultsData, 'participantCount', null);
-      const familiesCount = get(resultsData, 'familiesCountCardinality', null);
       const cohortIsEmpty = (!isLoading && !resultsData) || participantCount === 0;
-      const filesCardinality = get(resultsData, 'filesCardinality', 0);
 
-      const hasNoFile = filesCardinality === 0;
-      const hasNoCohortQuery = isEmpty(sqon.content);
-
-      const showDetailsHeader = () => (
-        <div className="cb-extra-actions-header">
-          {hasNoCohortQuery ? (
-            <H2>All Data</H2>
-          ) : (
-            <Fragment>
-              <H2>Cohort Results</H2>
-              <h3 className="cb-sub-heading" style={{ fontWeight: 'normal', marginRight: '10px' }}>
-                for Query {activeSqonIndex + 1}
-              </h3>
-            </Fragment>
-          )}
-        </div>
-      );
-
-      const extraActions = (
-        <div className="cb-extra-actions">
-          {showDetailsHeader()}
-          {isLoading ? (
-            <div className={'cb-summary-is-loading'}>
-              <Spin size={'small'} />
-            </div>
-          ) : (
-            <div className="cb-summary">
-              <div className="cb-summary-entity">
-                <h3 className="cb-sub-heading">
-                  <DemographicIcon width="14px" height="17px" />
-                  {formatCountResult(participantCount, 'participant')}
-                </h3>
-              </div>
-              <div className="cb-summary-entity">
-                <h3 className="cb-sub-heading">
-                  <img src={familyMembers} alt="" height="13px" />
-                  {formatCountResult(familiesCount, 'family')}
-                </h3>
-              </div>
-              <div className="cb-summary-entity">
-                {hasNoFile ? (
-                  <h3 className="cb-sub-heading-without-count">
-                    <div className="cb-summary-files">
-                      <div>
-                        <FilesIcon style={{ marginRight: '6px' }} /> {'No File'}
-                      </div>
-                    </div>
-                  </h3>
-                ) : (
-                  <div>
-                    <ButtonWithRouter
-                      getLink={
-                        hasNoCohortQuery
-                          ? () => '/search/file'
-                          : () => generateAllFilesLink(state.loggedInUser, api, sqon)
-                      }
-                    >
-                      {formatCountResult(filesCardinality, 'file')}
-                      <ArrowRightOutlined />
-                    </ButtonWithRouter>
-                  </div>
-                )}
-              </div>
-            </div>
-          )}
-        </div>
+      const toolbar = (
+        <Toolbar
+          {...{
+            isLoading,
+            data: data[0],
+            participantCount,
+            isFiltered,
+            activeSqonIndex,
+            egoGroups: state.egoGroups,
+            sqon,
+          }}
+        />
       );
 
       return (
         <Fragment>
           <div style={{ padding: '0 30px 0 34px' }} className="cb-view-links">
             <Tabs
-              tabBarExtraContent={extraActions}
+              tabBarExtraContent={toolbar}
               type="card"
               style={{ marginBottom: '0px' }}
               tabBarStyle={{ marginBottom: '0px' }}
@@ -258,7 +121,7 @@ const Results = ({ activeSqonIndex, sqon = { op: 'and', content: [] }, api, stat
                 className="cb-tab-content"
               >
                 {cohortIsEmpty ? <EmptyCohortOverlay /> : null}
-                <ParticipantsTableView sqon={sqon} egoGroups={state.egoGroups} />
+                <ParticipantsTableView sqon={sqon} />
               </TabPane>
             </Tabs>
           </div>