+++ src/uikit/DataTable/ToolbarButtons/ColumnFilter.js
@@ -1,9 +1,9 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { Button, Menu, Dropdown, Checkbox } from 'antd';
-import { DownOutlined } from '@ant-design/icons';
+import { ControlOutlined } from '@ant-design/icons';
 
-const filterCols = cols => cols.filter(col => col.filterable !== false);
+const filterCols = (cols) => cols.filter((col) => col.filterable !== false);
 
 class ColumnFilter extends Component {
   static defaultProps = {
@@ -20,10 +20,10 @@ class ColumnFilter extends Component {
 
   state = {
     isOpen: false,
-    checkedList: filterCols(this.props.defaultCols).map(i => i.Header),
+    checkedList: filterCols(this.props.defaultCols).map((i) => i.Header),
   };
 
-  onChangeCheckBox = e => {
+  onChangeCheckBox = (e) => {
     const { checkedList } = this.state;
     const { columns, onChange } = this.props;
 
@@ -31,11 +31,13 @@ class ColumnFilter extends Component {
     const isChecked = checkedList.includes(itemValue);
 
     const newCheckList = isChecked
-      ? checkedList.filter(i => i !== itemValue)
+      ? checkedList.filter((i) => i !== itemValue)
       : [...checkedList, itemValue];
 
-    const updatedCols = columns.map(c => (c.Header === itemValue ? { ...c, show: !isChecked } : c));
-    const updatedCol = updatedCols.find(c => c.Header === itemValue);
+    const updatedCols = columns.map((c) =>
+      c.Header === itemValue ? { ...c, show: !isChecked } : c,
+    );
+    const updatedCol = updatedCols.find((c) => c.Header === itemValue);
 
     onChange(updatedCols, updatedCol);
 
@@ -67,8 +69,13 @@ class ColumnFilter extends Component {
     const { colsPickerBtnClassName } = this.props;
     return (
       <Dropdown overlay={this.generateMenu()} trigger={['click']}>
-        <Button aria-label={`Show columns to select`} className={colsPickerBtnClassName}>
-          Columns <DownOutlined />
+        <Button
+          aria-label={`Show columns to select`}
+          className={colsPickerBtnClassName}
+          type="link"
+        >
+          <ControlOutlined />
+          Columns
         </Button>
       </Dropdown>
     );