+++ src/uikit/Headings.js
@@ -6,7 +6,6 @@ import styles from './Core.module.css';
  * Headings: Standard
  */
 export const H1 = styleComponent('h1', styles.heading1);
-export const H2 = styleComponent('h2', styles.heading2);
 export const H3 = styleComponent('h3', styles.heading3);
 export const H4 = styleComponent('h4', styles.heading4);